feat:Added countryDescription to SectionTwo
