+++ src/data/Nigeria/nigeriaData.js
@@ -1,0 +1,40 @@
+import Agriculture from '../../assets/images/Nigeria/SectionTwo/agriculture.webp';
+import Manufacturing from '../../assets/images/Nigeria/SectionTwo/manufacturing.webp';
+import Services from '../../assets/images/Nigeria/SectionTwo/services.webp';
+import Retail from '../../assets/images/Nigeria/SectionTwo/retail.webp';
+import ICT from '../../assets/images/Nigeria/SectionTwo/ICT.webp';
+
+const data = [
+  {
+    id: 1,
+    title: 'Agriculture',
+    image: Agriculture,
+    alt: 'Agriculture',
+  },
+  {
+    id: 2,
+    title: 'Manufacturing',
+    image: Manufacturing,
+    alt: 'Manufacturing',
+  },
+  {
+    id: 3,
+    title: 'Services',
+    image: Services,
+    alt: 'Services',
+  },
+  {
+    id: 4,
+    title: 'Retail',
+    image: Retail,
+    alt: 'Retail',
+  },
+  {
+    id: 5,
+    title: 'ICT',
+    image: ICT,
+    alt: 'ICT',
+  },
+];
+
+export default data;
