+++ src/data/Nigeria/nigeriaDesc.js
@@ -1,0 +1,28 @@
+const data = [
+  {
+    id: 1,
+    title: 'Land Area',
+    body: 'Nigeria offers a substantial land area of 923,769 sq. km.',
+    color: '#004F9F',
+  },
+  {
+    id: 2,
+    title: 'Population',
+    body: "As of 2023, Nigeria's population of 230.8 million promises a vast consumer base.",
+    color: '#007736',
+  },
+  {
+    id: 3,
+    title: 'GDP',
+    body: "Nigeria's GDP stood at US$400 Billion in 2022, indicating a significant economic footprint.",
+    color: '#E02626',
+  },
+  {
+    id: 4,
+    title: 'Labour Force',
+    body: 'With a labour force of 126 million in 2022, Nigeria boasts a substantial workforce.',
+    color: '#EF7D00',
+  },
+];
+
+export default data;
