--- src/data/nigeriaData.js
@@ -1,39 +1,0 @@
-import Agriculture from '../assets/images/Nigeria/SectionTwo/industries_agriculture.webp';
-import Manufacturing from '../assets/images/Nigeria/SectionTwo/industries_manufacturing.webp';
-const data = [
-  {
-    id: 1,
-    title: 'Agriculture',
-    image: Agriculture,
-    alt: 'Agriculture',
-  },
-  {
-    id: 2,
-    title: 'Manufacturing',
-    image: Manufacturing,
-    alt: 'Manufacturing',
-  },
-  {
-    id: 3,
-    title: 'Services',
-    image:
-      'https://images.unsplash.com/photo-1553484771-371a605b060b?q=80&w=1770&auto=format&fit=crop&ixlib=rb-4.0.3&ixid=M3wxMjA3fDB8MHxwaG90by1wYWdlfHx8fGVufDB8fHx8fA%3D%3D',
-    alt: 'Services',
-  },
-  {
-    id: 4,
-    title: 'Retail',
-    image:
-      'https://images.unsplash.com/photo-1525328437458-0c4d4db7cab4?q=80&w=1770&auto=format&fit=crop&ixlib=rb-4.0.3&ixid=M3wxMjA3fDB8MHxwaG90by1wYWdlfHx8fGVufDB8fHx8fA%3D%3D',
-    alt: 'Retail',
-  },
-  {
-    id: 5,
-    title: 'ICT',
-    image:
-      'https://images.unsplash.com/photo-1518770660439-4636190af475?q=80&w=1770&auto=format&fit=crop&ixlib=rb-4.0.3&ixid=M3wxMjA3fDB8MHxwaG90by1wYWdlfHx8fGVufDB8fHx8fA%3D%3D',
-    alt: 'ICT',
-  },
-];
-
-export default data;
