+++ src/pages/Nigeria/Nigeria.jsx
@@ -1,12 +1,14 @@
+import SectionTwo from './SectionTwo/SectionTwo';
 import SectionThree from './SectionThree/SectionThree';
-import SectionFour from './SectionFour/SectionFour';
+import Footer from '../../components/Footer/Footer';
 import styles from './Nigeria.module.css';
 
 function Nigeria() {
   return (
     <div className={styles.card_wrapper} id="nigeria" role="main">
+      <SectionTwo />
       <SectionThree />
-      <SectionFour />
+      <Footer />
     </div>
   );
 }
